refactor and fix flakes tests
