+++ app/src/androidTest/java/chat/rocket/android/chatdetails/ui/ChatDetailsFragmentTest.kt
@@ -3,7 +3,7 @@ package chat.rocket.android.chatdetails.ui
 import androidx.recyclerview.widget.RecyclerView
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.NoMatchingViewException
-import androidx.test.espresso.action.ViewActions.*
+import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.contrib.RecyclerViewActions
 import androidx.test.espresso.matcher.ViewMatchers.*
@@ -11,7 +11,8 @@ import androidx.test.rule.ActivityTestRule
 import chat.rocket.android.R
 import chat.rocket.android.analytics.event.ScreenViewEvent
 import chat.rocket.android.authentication.ui.AuthenticationActivity
-import chat.rocket.android.matchers.withTextInChild
+import chat.rocket.android.util.loginUserToTheApp
+import chat.rocket.android.util.withTextInChild
 import chat.rocket.android.util.extensions.addFragmentBackStack
 import org.junit.Before
 import org.junit.Rule
@@ -22,10 +23,8 @@ import testConfig.Config.Companion.FAVORITE_MESSAGES
 import testConfig.Config.Companion.FILES
 import testConfig.Config.Companion.MEMBERS
 import testConfig.Config.Companion.MENTIONS
-import testConfig.Config.Companion.PASSWORD
 import testConfig.Config.Companion.PINNED_MESSAGES
 import testConfig.Config.Companion.SERVER_URL
-import testConfig.Config.Companion.USERNAME
 
 class ChatDetailsFragmentTest {
 
@@ -38,7 +37,10 @@ class ChatDetailsFragmentTest {
     @Before
     fun setUp() {
         try {
-            loginIfUserIsLoggedOut()
+            rule().activity.addFragmentBackStack(ScreenViewEvent.Login.screenName, R.id.fragment_container) {
+                chat.rocket.android.authentication.login.ui.newInstance(SERVER_URL)
+            }
+            loginUserToTheApp()
         } catch (e: NoMatchingViewException) {
             Thread.sleep(3000)
         }
@@ -116,16 +118,6 @@ class ChatDetailsFragmentTest {
         navigateToExistingChannelDetails()
         onView(withText(PINNED_MESSAGES)).perform(click())
         onView(withId(R.id.pinned_messages_layout)).check(matches(isDisplayed()))
-    }
-
-    private fun loginIfUserIsLoggedOut() {
-        rule().activity.addFragmentBackStack(ScreenViewEvent.Login.screenName, R.id.fragment_container) {
-            chat.rocket.android.authentication.login.ui.newInstance(SERVER_URL)
-        }
-        onView(withId(R.id.text_username_or_email)).perform(typeText(USERNAME), closeSoftKeyboard())
-        onView(withId(R.id.text_password)).perform(typeText(PASSWORD), closeSoftKeyboard())
-        onView(withId(R.id.button_log_in)).perform(click())
-        Thread.sleep(12000)
     }
 
     private fun navigateToExistingDMDetails() {
