+++ app/src/androidTest/java/chat/rocket/android/favoritemessages/ui/FavoriteMessagesFragmentTest.kt
@@ -9,7 +9,8 @@ import androidx.test.rule.ActivityTestRule
 import chat.rocket.android.R
 import chat.rocket.android.analytics.event.ScreenViewEvent
 import chat.rocket.android.authentication.ui.AuthenticationActivity
-import chat.rocket.android.matchers.RecyclerViewItemCountAssertion.Companion.withItemCount
+import chat.rocket.android.util.RecyclerViewItemCountAssertion.Companion.withItemCount
+import chat.rocket.android.util.loginUserToTheApp
 import chat.rocket.android.util.extensions.addFragmentBackStack
 import org.hamcrest.Matchers.equalTo
 import org.hamcrest.Matchers.greaterThan
@@ -17,9 +18,7 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import testConfig.Config.Companion.FAVORITE_MESSAGES
-import testConfig.Config.Companion.PASSWORD
 import testConfig.Config.Companion.SERVER_URL
-import testConfig.Config.Companion.USERNAME
 
 
 class FavoriteMessagesFragmentTest {
@@ -33,7 +32,10 @@ class FavoriteMessagesFragmentTest {
     @Before
     fun setUp() {
         try {
-            loginIfUserIsLoggedOut()
+            rule().activity.addFragmentBackStack(ScreenViewEvent.Login.screenName, R.id.fragment_container) {
+                chat.rocket.android.authentication.login.ui.newInstance(SERVER_URL)
+            }
+            loginUserToTheApp()
         } catch (e: NoMatchingViewException) {
         }
     }
@@ -52,20 +54,6 @@ class FavoriteMessagesFragmentTest {
         onView(withText(FAVORITE_MESSAGES)).perform(click())
         Thread.sleep(6000)
         onView(withId(R.id.recycler_view)).check(withItemCount(equalTo(0)))
-    }
-
-
-    private fun loginIfUserIsLoggedOut(){
-        rule().activity.addFragmentBackStack(ScreenViewEvent.Login.screenName, R.id.fragment_container) {
-            chat.rocket.android.authentication.login.ui.newInstance(SERVER_URL)
-        }
-        onView(withId(R.id.text_username_or_email)).perform(
-            typeText(USERNAME),
-            closeSoftKeyboard()
-        )
-        onView(withId(R.id.text_password)).perform(typeText(PASSWORD), closeSoftKeyboard())
-        onView(withId(R.id.button_log_in)).perform(click())
-        Thread.sleep(12000)
     }
 
     private fun navigateToSandboxChannelDetails() {
