--- app/src/androidTest/java/chat/rocket/android/matchers/RecyclerViewCountAssertion.kt
@@ -1,29 +1,0 @@
-package chat.rocket.android.matchers
-
-import android.view.View
-import androidx.recyclerview.widget.RecyclerView
-import androidx.test.espresso.NoMatchingViewException
-import androidx.test.espresso.ViewAssertion
-import org.hamcrest.Matcher
-import org.hamcrest.MatcherAssert.assertThat
-
-
-class RecyclerViewItemCountAssertion private constructor(private val matcher: Matcher<Int>) :
-    ViewAssertion {
-
-    override fun check(view: View, noViewFoundException: NoMatchingViewException?) {
-        if (noViewFoundException != null) {
-            throw noViewFoundException
-        }
-        val recyclerView = view as RecyclerView
-        val adapter = recyclerView.adapter
-        assertThat(adapter!!.itemCount, matcher)
-    }
-
-    companion object {
-
-        fun withItemCount(matcher: Matcher<Int>): RecyclerViewItemCountAssertion {
-            return RecyclerViewItemCountAssertion(matcher)
-        }
-    }
-}
