+++ app/src/androidTest/java/chat/rocket/android/members/ui/MembersFragmentTest.kt
@@ -2,14 +2,15 @@ package chat.rocket.android.members.ui
 
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.NoMatchingViewException
-import androidx.test.espresso.action.ViewActions.*
+import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.rule.ActivityTestRule
 import chat.rocket.android.R
 import chat.rocket.android.analytics.event.ScreenViewEvent
 import chat.rocket.android.authentication.ui.AuthenticationActivity
-import chat.rocket.android.matchers.RecyclerViewItemCountAssertion.Companion.withItemCount
+import chat.rocket.android.util.RecyclerViewItemCountAssertion.Companion.withItemCount
+import chat.rocket.android.util.loginUserToTheApp
 import chat.rocket.android.util.extensions.addFragmentBackStack
 import org.hamcrest.Matchers.greaterThan
 import org.junit.Before
@@ -17,9 +18,7 @@ import org.junit.Rule
 import org.junit.Test
 import testConfig.Config.Companion.EXISTING_CHANNEL
 import testConfig.Config.Companion.MEMBERS
-import testConfig.Config.Companion.PASSWORD
 import testConfig.Config.Companion.SERVER_URL
-import testConfig.Config.Companion.USERNAME
 
 
 class MembersFragmentTest {
@@ -33,7 +32,10 @@ class MembersFragmentTest {
     @Before
     fun setUp() {
         try {
-            loginIfUserIsLoggedOut()
+            rule().activity.addFragmentBackStack(ScreenViewEvent.Login.screenName, R.id.fragment_container) {
+                chat.rocket.android.authentication.login.ui.newInstance(SERVER_URL)
+            }
+            loginUserToTheApp()
             navigateToChannelDetails()
         } catch (e: NoMatchingViewException) {
             Thread.sleep(3000)
@@ -46,19 +48,6 @@ class MembersFragmentTest {
         onView(withText(MEMBERS)).perform(click())
         Thread.sleep(6000)
         onView(withId(R.id.recycler_view)).check(withItemCount(greaterThan(0)))
-    }
-
-    private fun loginIfUserIsLoggedOut(){
-        rule().activity.addFragmentBackStack(ScreenViewEvent.Login.screenName, R.id.fragment_container) {
-            chat.rocket.android.authentication.login.ui.newInstance(SERVER_URL)
-        }
-        onView(withId(R.id.text_username_or_email)).perform(
-            typeText(USERNAME),
-            closeSoftKeyboard()
-        )
-        onView(withId(R.id.text_password)).perform(typeText(PASSWORD), closeSoftKeyboard())
-        onView(withId(R.id.button_log_in)).perform(click())
-        Thread.sleep(12000)
     }
 
     private fun navigateToChannelDetails() {
