+++ app/src/androidTest/java/chat/rocket/android/profile/ui/ProfileFragmentTest.kt
@@ -2,7 +2,7 @@ package chat.rocket.android.profile.ui
 
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.NoMatchingViewException
-import androidx.test.espresso.action.ViewActions.*
+import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.filters.LargeTest
@@ -10,6 +10,7 @@ import androidx.test.rule.ActivityTestRule
 import chat.rocket.android.R
 import chat.rocket.android.analytics.event.ScreenViewEvent
 import chat.rocket.android.authentication.ui.AuthenticationActivity
+import chat.rocket.android.util.loginUserToTheApp
 import chat.rocket.android.util.extensions.addFragmentBackStack
 import org.junit.Before
 import org.junit.Rule
@@ -21,7 +22,6 @@ import testConfig.Config.Companion.EMAIL
 import testConfig.Config.Companion.INVISIBLE
 import testConfig.Config.Companion.NAME
 import testConfig.Config.Companion.ONLINE
-import testConfig.Config.Companion.PASSWORD
 import testConfig.Config.Companion.SERVER_URL
 import testConfig.Config.Companion.USERNAME
 
@@ -37,7 +37,10 @@ class ProfileFragmentTest {
     @Before
     fun setUp() {
         try {
-            loginIfUserIsLoggedOut()
+            rule().activity.addFragmentBackStack(ScreenViewEvent.Login.screenName, R.id.fragment_container) {
+                chat.rocket.android.authentication.login.ui.newInstance(SERVER_URL)
+            }
+            loginUserToTheApp()
             navigateToProfileFragment()
         } catch (e: NoMatchingViewException) {
             Thread.sleep(4000)
@@ -102,18 +105,6 @@ class ProfileFragmentTest {
         onView(withText(INVISIBLE)).perform(click())
         onView(withText(CHANGE_STATUS)).perform(click())
         onView(withId(R.id.text_status)).check(matches(withText("Status: Offline")))
-    }
-
-    private fun loginIfUserIsLoggedOut() {
-        rule().activity.addFragmentBackStack(ScreenViewEvent.Login.screenName, R.id.fragment_container) {
-            chat.rocket.android.authentication.login.ui.newInstance(SERVER_URL)
-        }
-        onView(withId(R.id.text_username_or_email)).perform(
-            typeText(USERNAME), closeSoftKeyboard()
-        )
-        onView(withId(R.id.text_password)).perform(typeText(PASSWORD), closeSoftKeyboard())
-        onView(withId(R.id.button_log_in)).perform(click())
-        Thread.sleep(12000)
     }
 
     private fun navigateToProfileFragment() {
