+++ app/src/androidTest/java/chat/rocket/android/servers/ui/ServersBottomSheetFragmentTest.kt
@@ -2,25 +2,22 @@ package chat.rocket.android.servers.ui
 
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.NoMatchingViewException
-import androidx.test.espresso.action.ViewActions.*
-import androidx.test.espresso.assertion.ViewAssertions
+import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.rule.ActivityTestRule
 import chat.rocket.android.R
 import chat.rocket.android.analytics.event.ScreenViewEvent
 import chat.rocket.android.authentication.ui.AuthenticationActivity
-import chat.rocket.android.matchers.RecyclerViewItemCountAssertion
+import chat.rocket.android.util.RecyclerViewItemCountAssertion
+import chat.rocket.android.util.loginUserToTheApp
 import chat.rocket.android.util.extensions.addFragmentBackStack
 import org.hamcrest.Matchers
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import testConfig.Config.Companion.ORG_NAME
-import testConfig.Config.Companion.PASSWORD
 import testConfig.Config.Companion.SERVER_URL
-import testConfig.Config.Companion.USERNAME
 
 class ServersBottomSheetFragmentTest {
 
@@ -33,25 +30,15 @@ class ServersBottomSheetFragmentTest {
     @Before
     fun setUp() {
         try {
-            loginIfUserIsLoggedOut()
+            rule().activity.addFragmentBackStack(ScreenViewEvent.Login.screenName, R.id.fragment_container) {
+                chat.rocket.android.authentication.login.ui.newInstance(SERVER_URL)
+            }
+            loginUserToTheApp()
             onView(withText(ORG_NAME)).perform(click())
         } catch (e: NoMatchingViewException) {
             Thread.sleep(3000)
             onView(withText(ORG_NAME)).perform(click())
         }
-    }
-
-    private fun loginIfUserIsLoggedOut(){
-        rule().activity.addFragmentBackStack(ScreenViewEvent.Login.screenName, R.id.fragment_container) {
-            chat.rocket.android.authentication.login.ui.newInstance(SERVER_URL)
-        }
-        onView(withId(R.id.text_username_or_email)).perform(
-            typeText(USERNAME),
-            closeSoftKeyboard()
-        )
-        onView(withId(R.id.text_password)).perform(typeText(PASSWORD), closeSoftKeyboard())
-        onView(withId(R.id.button_log_in)).perform(click())
-        Thread.sleep(12000)
     }
 
     @Test
