+++ app/src/androidTest/java/chat/rocket/android/userdetails/ui/UserDetailsFragmentTest.kt
@@ -10,11 +10,11 @@ import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.rule.ActivityTestRule
 import chat.rocket.android.R
 import chat.rocket.android.main.ui.MainActivity
-import chat.rocket.android.matchers.clickChildViewWithId
-import kotlinx.android.synthetic.main.fragment_chat_rooms.*
+import chat.rocket.android.util.clickChildViewWithId
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
+import testConfig.Config.Companion.EXISTING_USER2
 
 class UserDetailsFragmentTest {
 
@@ -27,12 +27,7 @@ class UserDetailsFragmentTest {
     @Before
     fun setUp() {
         Thread.sleep(5000)
-        onView(withId(R.id.recycler_view))
-            .perform(
-                RecyclerViewActions.actionOnItemAtPosition<RecyclerView.ViewHolder>(
-                    activityRule.activity.recycler_view.adapter!!.itemCount - 1, click()
-                )
-            )
+        onView(withText(EXISTING_USER2)).perform(click())
         Thread.sleep(5000)
         onView(withId(R.id.recycler_view)).perform(
             RecyclerViewActions.actionOnItemAtPosition<RecyclerView.ViewHolder>(
@@ -40,21 +35,21 @@ class UserDetailsFragmentTest {
                 clickChildViewWithId(R.id.image_avatar)
             )
         )
-        Thread.sleep(5000)
+        Thread.sleep(2000)
     }
 
     @Test
     fun check_UI_elements() {
+        Thread.sleep(5000)
         onView(withId(R.id.image_avatar)).check(matches(isDisplayed()))
         onView(withId(R.id.image_blur)).check(matches(isDisplayed()))
         onView(withId(R.id.image_arrow_back)).check(matches(isDisplayed()))
-        onView(withId(R.id.text_name)).check(matches(isDisplayed()))
-        onView(withId(R.id.text_username)).check(matches(isDisplayed()))
         onView(withId(R.id.text_message)).check(matches(withText("Message")))
         onView(withId(R.id.text_video_call)).check(matches(withText("Video call")))
         onView(withId(R.id.text_title_status)).check(matches(withText("Status")))
         onView(withId(R.id.text_title_timezone)).check(matches(withText("Timezone")))
         onView(withId(R.id.text_description_timezone)).check(matches(isDisplayed()))
+        onView(withId(R.id.button_remove_user)).check(matches(isDisplayed()))
     }
 
     @Test
